homebuy_ai/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pyyaml==6.0.2
pydantic==2.8.2
feedparser==6.0.11
requests==2.32.3
python-dateutil==2.9.0.post0
plotly==5.23.0
streamlit==1.37.1
openai==1.40.3
